refactor: refactor to a more service-oriented architecture
