+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const FUNDS_RANGE_START = 20;
+export const FUNDS_RANGE_END = 500;
